# Improve adding to a dense matrix