@MichielStock 2019-09-05
OK thanks. Originally, the idea was that addition is not supported since this can only be done using the dense matrices. Since `getindex` apparently provides a method for this automatically, it makes sense to provide the faster alternative.